requirements.txt
# PyTorch

torch
pytorch-lightning
torchvision

# Version numbers have been retrieved from a range of machines and environments.
# Take them with a grain of salt.

# Direct dependencies
#python
#pip
#setuptools
numpy
pandas
matplotlib
seaborn
scikit-learn #
librosa
audioread
mutagen
pydub
#exiftool  # Only considered at some point.
#eyed3  # Only considered at some point.
requests
pydot
tqdm
python-dotenv

# # Dependencies of the above.
# certifi
# click
# cycler
# Cython
# decorator
# joblib
# protobuf
# pyparsing
# python-dateutil
# pytz
# PyYAML
# resampy  # 
# scipy #
# six
# Theano

# # Jupyter notebook and its dependencies.
# notebook
# ipywidgets
# bleach
# entrypoints
# html5lib
# ipykernel
# ipython
# ipython-genutils
# Jinja2
# jsonschema
# jupyter-client
# jupyter-core
# MarkupSafe
# mistune
# nbconvert
# nbformat
# packaging
# pandocfilters
# pexpect
# pickleshare
# prompt-toolkit
# ptyprocess
# Pygments
# pyparsing
# pyzmq
# simplegeneric
# terminado
# testpath
# tornado
# traitlets
# wcwidth
# webencodings
# widgetsnbextension
